Update deprecated use of `crypto:hmac_` functions
In R23 the `crypto:hmac_` functions are deprecated and will be removed in R24.
In this patch we're introducing a compatability layer so that when compiled
with R23 or newer, where the "new" crypto API was introduced, the `crypto:mac`
functions will be used instead.

When creating this patch I noticed that the call to `crypto:hmac/3` had been
replaced in 267eb2c97c07f7f5bec83c12b36cd6cd3a90d5e5 with a chain of calls to
the version of the functions which update with dynamic data.  It's wasn't clear
to me why that change was made so I didn't undo it, but I have used the single
call in the updated function so create the HMAC in a single step.