Fix: show proper calling arguments in README example
In the README there are a couple examples of calling `pbkdf2` which are missing
the first argument, the mac function type. There are a few different ways to
call into this library but all of them require the mac function type as the
first argument.

In this patch we're updating the README to reflect the proper usage of the function.